Added fuctionality to add order and picture for model's

+++ Customer.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 
 string Customer::getName(){ return name; }
-double Customer::getCustomerNum(){ return customerNum; }
+int Customer::getCustomerNum(){ return customerNum; }
 double Customer::getWallet(){ return wallet; }
 void Customer::setWallet(double newWallet){ wallet = newWallet; }
 //void Customer::addOrder(Order order){ orders.push_back(order); }
